+++ internal/fsutil/fsutil.go
@@ -1,0 +1,21 @@
+package fsutil
+
+import (
+	"os"
+	"path/filepath"
+)
+
+// CalculateDirSize calculates the total size of all files in a directory recursively.
+func CalculateDirSize(path string) (int64, error) {
+	var size int64
+	err := filepath.Walk(path, func(_ string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+		if !info.IsDir() {
+			size += info.Size()
+		}
+		return nil
+	})
+	return size, err
+}